Testing Queries
A file to test SQL Queries from PHP with permission granted from my database. Insert does NOT want to work, although deleting and displaying works fine. I believe it is a problem with the form, as it it not receiving the values from it. The login and grants refer to my database.